+++ winter-android-integration-test/src/main/java/io/jentz/winter/android/test/viewmodel/ViewModel.kt
@@ -1,0 +1,7 @@
+package io.jentz.winter.android.test.viewmodel
+
+import io.reactivex.Flowable
+
+interface ViewModel<VS> {
+    fun toFlowable(): Flowable<VS>
+}